requirements_cpu.txt
absl-py==0.11.0
appdirs==1.4.4
astunparse==1.6.3
attrs==20.3.0
cached-property==1.5.2
cachetools==4.2.0
certifi==2020.12.5
chardet==3.0.4
cycler==0.10.0
defusedxml==0.7.0rc1
gast==0.3.3
google-auth==1.24.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.34.0
h5py==2.10.0
idna==2.10
importlib-metadata==3.1.1
imutils==0.5.3
isodate==0.6.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
lxml==4.6.2
Markdown==3.3.3
matplotlib==3.3.3
numpy==1.18.5
oauthlib==3.1.0
onvif-zeep==0.2.12
opt-einsum==3.3.0
Pillow==8.0.1
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.4
requests==2.25.0
requests-file==1.5.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rsa==4.6
six==1.15.0
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.1
tensorflow-estimator==2.3.0
termcolor==1.1.0
urllib3==1.26.2
Werkzeug==1.0.1
wrapt==1.12.1
zeep==4.0.0
zipp==3.4.0
scipy==1.6.3
